module7-task3: Rename film store to films
