+++ project/src/components/layout/film-card-list/film-card-list.tsx
@@ -3,7 +3,7 @@ import SmallFilmCard from '../small-film-card/small-film-card';
 
 type Props = {
   films: Array<Film>;
-}
+};
 
 function FilmCardList({ films }: Props): JSX.Element {
   return (
