+++ project/src/components/layout/genre-list/genre-list.tsx
@@ -1,5 +1,5 @@
-import { changeGenre, resetFilmCardsCount } from '@store/film/actions';
-import { getActiveGenre, getGenres } from '@store/film/selectors';
+import { changeGenre, resetFilmCardsCount } from '@store/films/actions';
+import { getActiveGenre, getGenres } from '@store/films/selectors';
 import { useDispatch, useSelector } from 'react-redux';
 
 function GenreList(): JSX.Element {
