+++ project/src/components/layout/similar-film-card-list/similar-film-card-list.tsx
@@ -1,4 +1,4 @@
-import { getSimilarFilmsFactory } from '@store/film/selectors';
+import { getSimilarFilmsFactory } from '@store/films/selectors';
 import { useSelector } from 'react-redux';
 import { useParams } from 'react-router';
 import { RouteParams } from 'src/types/route-params';
