+++ project/src/components/pages/film/film.tsx
@@ -11,7 +11,7 @@ import Reviews from '@components/layout/reviews/reviews';
 import Tabs, { TabConfig } from '@components/layout/tabs/tabs';
 import SimilarFilmCardList from '@components/layout/similar-film-card-list/similar-film-card-list';
 import { useSelector } from 'react-redux';
-import { getFilms } from '@store/film/selectors';
+import { getFilms } from '@store/films/selectors';
 import Poster from '@components/layout/poster/poster';
 import FilmCardButtons from '@components/layout/film-card-buttons/film-card-buttons';
 import FilmCardHeader from '@components/layout/film-card-header/film-card-header';
