+++ project/src/index.tsx
@@ -4,12 +4,12 @@ import { applyMiddleware, combineReducers, createStore } from 'redux';
 import { Provider } from 'react-redux';
 import App from './components/app/app';
 import { PROMO_FILM_MOCK } from './mocks/promo-film';
-import { reducer as filmReducer } from '@store/film/reducer';
+import { reducer as filmReducer } from '@store/films/reducer';
 import { reducer as userReducer } from '@store/user/reducer';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import { createAPI } from './api/api';
 import thunk from 'redux-thunk';
-import { getFilms } from './store/film/async-actions';
+import { getFilms } from './store/films/async-actions';
 import { ThunkApiDispatch } from './types/actions';
 import { GlobalState } from './types/global-state';
 import { checkAuth } from '@store/user/async-actions';
@@ -22,7 +22,7 @@ const api = createAPI(() => {
 });
 
 const reducer = combineReducers<GlobalState>({
-  film: filmReducer,
+  films: filmReducer,
   user: userReducer,
 });
