+++ project/src/store/films/reducer.ts
@@ -1,16 +1,16 @@
 import { Reducer } from 'redux';
 import { DEFALUT_ACTIVE_GENRE, FILM_CARDS_COUNT } from 'src/constants';
 import { Actions, ActionType } from 'src/types/actions';
-import { FilmState } from 'src/types/global-state';
+import { FilmsState } from 'src/types/global-state';
 
-const initialState: FilmState = {
+const initialState: FilmsState = {
   films: [],
   filmsLoaded: false,
   genre: DEFALUT_ACTIVE_GENRE,
   filmCardsCount: FILM_CARDS_COUNT,
 };
 
-export const reducer: Reducer<FilmState, Actions> = (state = initialState,action): FilmState => {
+export const reducer: Reducer<FilmsState, Actions> = (state = initialState, action): FilmsState => {
   switch (action.type) {
     case ActionType.ChangeGenre:
       return { ...state, genre: action.payload };
